Fix: TIL도 뷰모델 사용하게 변경 + 메인 네비게이션 추가
