+++ ORI/View/Main/TILListViewController.swift
@@ -8,38 +8,30 @@
 import UIKit
 
 class TILListViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    var posts: [TIL] = []
-    let tableView = UITableView()
+    let viewModel = TILListViewModel()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        mainNavigationBar()
         setupTableView()
-        fetchTILList(page: 0, size: 1) { [weak self] result in
-            switch result {
-            case .success(let posts):
-                self?.posts = posts
-                self?.tableView.reloadData()
-            case .failure(let error):
-                print("Error fetching posts: \(error)")
-            }
-        }
+        viewModel.loadTILList()
     }
     
     func setupTableView() {
-        tableView.frame = view.bounds
-        tableView.dataSource = self
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "PostCell")
-        view.addSubview(tableView)
+        viewModel.tableView.frame = view.bounds
+        viewModel.tableView.dataSource = self
+        viewModel.tableView.register(UITableViewCell.self, forCellReuseIdentifier: "PostCell")
+        view.addSubview(viewModel.tableView)
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return posts.count
+        return viewModel.posts.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "PostCell", for: indexPath)
-        let post = posts[indexPath.row]
+        let post = viewModel.posts[indexPath.row]
         cell.textLabel?.text = post.title
         return cell
     }
