+++ ORI/ViewModel/TILListViewModel.swift
@@ -6,7 +6,21 @@
 //
 
 import UIKit
+import UIKit
 
 class TILListViewModel {
+    var posts: [TIL] = []
+    let tableView = UITableView()
     
+    func loadTILList() {
+        fetchTILList(page: 0, size: 10) { [weak self] result in
+            switch result {
+            case .success(let posts):
+                self?.posts = posts
+                self?.tableView.reloadData()
+            case .failure(let error):
+                print("Error fetching posts: \(error)")
+            }
+        }
+    }
 }
